# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sale Management API",
    "version": "1.0.0",
    "description": "API documentation for Sale Management System"
  },
  "servers": [
    { "url": "http://localhost:4000" }
  ],
  "paths": {
...
    "/api/jobs": {
      "get": { "summary": "Get all jobs", "responses": { "200": { "description": "List of jobs" } } },
      "post": { "summary": "Create a job", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Job" } } } }, "responses": { "200": { "description": "Job created" } } }
    },
    "/api/jobs/{id}": {
      "get": { "summary": "Get job by ID", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Job data" } } },
      "put": { "summary": "Update job", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Job" } } } }, "responses": { "200": { "description": "Job updated" } } },
      "delete": { "summary": "Delete job", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Job deleted" } } }
    },
...
    "/api/product-types": {
      "get": { "summary": "Get all product types", "responses": { "200": { "description": "List of product types" } } },
      "post": { "summary": "Create a product type", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductType" } } } }, "responses": { "200": { "description": "Product type created" } } }
    },
    "/api/product-types/{id}": {
      "get": { "summary": "Get product type by ID", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Product type data" } } },
      "put": { "summary": "Update product type", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductType" } } } }, "responses": { "200": { "description": "Product type updated" } } },
      "delete": { "summary": "Delete product type", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Product type deleted" } } }
    },
    "/api/products": {
      "get": { "summary": "Get all products", "responses": { "200": { "description": "List of products" } } },
      "post": { "summary": "Create a product", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } }, "responses": { "200": { "description": "Product created" } } }
    },
    "/api/products/{id}": {
      "get": { "summary": "Get product by ID", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Product data" } } },
      "put": { "summary": "Update product", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } }, "responses": { "200": { "description": "Product updated" } } },
      "delete": { "summary": "Delete product", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Product deleted" } } }
    },
    "/api/invoices": {
      "get": { "summary": "Get all invoices", "responses": { "200": { "description": "List of invoices" } } },
      "post": { "summary": "Create an invoice", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Invoice" } } } }, "responses": { "200": { "description": "Invoice created" } } }
    },
    "/api/invoices/{id}": {
      "get": { "summary": "Get invoice by ID", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Invoice data" } } },
      "put": { "summary": "Update invoice", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Invoice" } } } }, "responses": { "200": { "description": "Invoice updated" } } },
      "delete": { "summary": "Delete invoice", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Invoice deleted" } } }
    },
...
  },
  "components": {
    "schemas": {
...
      "Job": { "type": "object", "properties": { "Job_Title": { "type": "string" }, "Department": { "type": "string" }, "Min_Salary": { "type": "number" }, "Max_Salary": { "type": "number" } }, "required": ["Job_Title", "Department"] },
...
      "ProductType": { "type": "object", "properties": { "ProductType_ID": { "type": "string" }, "ProductType_Name": { "type": "string" }, "Remarks": { "type": "string" } }, "required": ["ProductType_ID", "ProductType_Name"] },
      "Product": { "type": "object", "properties": { "Product_no": { "type": "string" }, "ProductName": { "type": "string" }, "ProductType": { "type": "string" }, "Profit_percent": { "type": "number" }, "Unit_measure": { "type": "string" }, "Reorder_level": { "type": "number" }, "Cost_Price": { "type": "number" }, "Sell_price": { "type": "number" }, "QTY_ON_HAND": { "type": "number" }, "Photo": { "type": "string", "format": "binary" } }, "required": ["Product_no", "ProductName"] },
      "Invoice": { "type": "object", "properties": { "Client_no": { "type": "integer" }, "EmployeeID": { "type": "integer" }, "Invoice_status": { "type": "string" }, "InvoiceMemo": { "type": "string" }, "Invoice_date": { "type": "string", "format": "date" } }, "required": ["Client_no", "EmployeeID"] },
...
    }
  }
}